JEAZINDWRK-25 Finished Milestone 2 Assignment DVD Library

+++ netBeans Java Projects/softwareguild/MilestoneTwo/ExerciseAddressBook/src/main/java/com/sg/addressbook/controller/AddressController.java
@@ -6,6 +6,7 @@
 package com.sg.addressbook.controller;
 
 import com.sg.addressbook.dao.AddressBookDao;
+import com.sg.addressbook.dao.AddressBookDaoException;
 import com.sg.addressbook.dto.Address;
 import com.sg.addressbook.ui.AddressBookView;
 import java.util.List;
@@ -26,7 +27,9 @@ public class AddressController {
     public void run() {
         boolean keepgoing = true;
         int menuSelection = 0;
-        while (keepgoing) {
+        
+        try {
+           while (keepgoing) {
             menuSelection = view.printMenuSelection();
             
             switch (menuSelection) {
@@ -40,22 +43,29 @@ public class AddressController {
                     findAddress();
                     break;
                 case 4:
-                    listAddressCount();
+                    editAddress();
                     break;
                 case 5:
-                    listAllAddresses();
+                    listAddressCount();
                     break;
                 case 6:
+                    listAllAddresses();
+                    break;
+                case 7:
                     keepgoing = false;
                     break;
                 default:
                     view.displayUnknownCommand();
             }        
+         }
+        view.displayExit(); 
+        } catch (AddressBookDaoException e) {
+            view.displayErrorMessage(e.getMessage());
         }
-        view.displayExit();
+        
     }
     
-    private void addNewAddress() {
+    private void addNewAddress() throws AddressBookDaoException {
         view.displayCreateNewAddress();
         Address newAddress = view.getNewAddress();
         dao.addAddress(newAddress.getLastName(), newAddress);
@@ -63,15 +73,21 @@ public class AddressController {
         view.displayContinue();
     }
     
-    private void deleteAddress() {
+    private void deleteAddress() throws AddressBookDaoException {
         view.displayDeletingAddress();
         String deletedAddressChoice = view.getAddress("delete");
-        dao.deleteAddress(deletedAddressChoice);
+        view.displayAddress(dao.findAddress(deletedAddressChoice));
+        boolean verification = view.verify("delete");
+        if (verification) {
+            dao.deleteAddress(deletedAddressChoice);
         view.displayDeletedAddressSuccess();
+        } else {
+            view.displayDeletedAddressAborted();
+        }
         view.displayContinue();
     }
     
-    private void findAddress() {
+    private void findAddress() throws AddressBookDaoException {
         view.displayFindingAddress();
         String FindAddressChoice = view.getAddress("find");
         Address foundAddress = dao.findAddress(FindAddressChoice);
@@ -79,14 +95,30 @@ public class AddressController {
         view.displayContinue();
     }
     
-    private void listAddressCount() {
+    private void editAddress() throws AddressBookDaoException {
+        view.displayEditingAddress();
+        String editAddressChoice = view.getAddress("edit");
+        Address addressToEdit = dao.findAddress(editAddressChoice);
+        view.displayAddress(addressToEdit);
+        boolean verification = view.verify("edit");
+        if (verification) {
+            Address editedAddress = view.editAddress(addressToEdit);
+            dao.addAddress(editedAddress.getLastName(), editedAddress);
+            view.displayEditedAddressSuccess();
+        } else {
+            view.displayEditAddressAborted();
+        }
+        view.displayContinue();
+    }
+    
+    private void listAddressCount() throws AddressBookDaoException {
         view.displayListingAddressCount();
         int addressCount = dao.listAddressCount();
         view.displayAddressCount(addressCount);
         view.displayContinue();
     }
     
-    private void listAllAddresses() {
+    private void listAllAddresses() throws AddressBookDaoException{
         view.displayListingAllAddresses();
         List<Address> allAddresses = dao.listAddresses();
         view.displayAllAddresses(allAddresses);
